+++ plugin_ide.core/src-lang/melnorme/lang/ide/core/utils/operation/CommonOperationCallable.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015, 2015 IBM Corporation and others.
+ * Copyright (c) 2015 Bruno Medeiros and other Contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,18 +8,18 @@
  * Contributors:
  *     Bruno Medeiros - initial API and implementation
  *******************************************************************************/
-package com.github.rustdt.ide.core.operations;
+package melnorme.lang.ide.core.utils.operation;
 
-import com.github.rustdt.tooling.ops.RustSDKLocationValidator;
+import java.util.concurrent.Callable;
 
-import melnorme.lang.ide.core.operations.AbstractToolManager;
-import melnorme.lang.tooling.ops.util.PathValidator;
+import org.eclipse.core.runtime.CoreException;
 
-public class RustToolManager extends AbstractToolManager {
+import melnorme.utilbox.concurrency.OperationCancellation;
+import melnorme.utilbox.core.CommonException;
+
+public interface CommonOperationCallable<RET> extends Callable<RET> {
 	
 	@Override
-	public PathValidator getSDKToolPathValidator() {
-		return new RustSDKLocationValidator();
-	}
+	RET call() throws CoreException, CommonException, OperationCancellation;
 	
 }