+++ plugin_ide.core/src-lang/melnorme/lang/ide/core/utils/operation/CommonProgressRunnable.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015, 2015 IBM Corporation and others.
+ * Copyright (c) 2015 Bruno Medeiros and other Contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,18 +8,15 @@
  * Contributors:
  *     Bruno Medeiros - initial API and implementation
  *******************************************************************************/
-package com.github.rustdt.ide.core.operations;
+package melnorme.lang.ide.core.utils.operation;
 
-import com.github.rustdt.tooling.ops.RustSDKLocationValidator;
+import org.eclipse.core.runtime.IProgressMonitor;
 
-import melnorme.lang.ide.core.operations.AbstractToolManager;
-import melnorme.lang.tooling.ops.util.PathValidator;
+import melnorme.utilbox.concurrency.OperationCancellation;
+import melnorme.utilbox.core.CommonException;
 
-public class RustToolManager extends AbstractToolManager {
+public interface CommonProgressRunnable {
 	
-	@Override
-	public PathValidator getSDKToolPathValidator() {
-		return new RustSDKLocationValidator();
-	}
+	void run(IProgressMonitor pm) throws CommonException, OperationCancellation;
 	
 }