+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Added Download operation to Racer preferences field (uses `cargo install`). FIXME: Guide, Source 
  * Fixed regression: Console view always activates when a build is invoked.
 
 ### 0.4.2